* changes:
  camera: replace `-Wno-error` flag with `-Werror`
  camera: fix `-Wunused-function` compiler warnings
  camera: fix `-Wpessimizing-move` compiler warnings
  camera: fix `-Wmissing-field-initializers` compiler warnings
  camera: fix `-Wsign-compare` compiler warnings
  camera: fix `-Wunused-variable` compiler warnings
  camera: fix `-Wunused-private-field` compiler warnings
  camera: fix `-Wreorder` compiler warnings
  camera: fix `-Wunused-parameter` compiler warnings
  camera: fix `-Wformat` compiler warnings
  camera: fix `-Wmacro-redefined` compiler warnings

+++ modules/camera/3_4/metadata/metadata_reader.cpp
@@ -14,10 +14,11 @@
  * limitations under the License.
  */
 
-#include "metadata_reader.h"
-
 // #define LOG_NDEBUG 0
 #define LOG_TAG "MetadataReader"
+
+#include "metadata_reader.h"
+
 #include <cutils/log.h>
 #include <system/camera.h>
 
@@ -200,9 +201,9 @@ int MetadataReader::StreamStallDurations(
     }
     // Must have a non-negative stall.
     if (stall.duration < 0) {
-      ALOGE("%s: Invalid stall duration: negative stall %d.",
+      ALOGE("%s: Invalid stall duration: negative stall %lld.",
             __func__,
-            stall.duration);
+            static_cast<long long>(stall.duration));
       return -EINVAL;
     }
     // TODO(b/31384253): YUV_420_888, RAW10, RAW12, RAW_OPAQUE,
